Baching is compiling, still not working

+++ Light-OpenGL/Resources/Shaders/BatchQuads.shader
@@ -5,10 +5,10 @@ layout(location = 0) in vec4 position;
 
 layout(location = 1) in vec2 texCoord;
 layout(location = 2) in vec4 color;
-layout(location = 3) in int texID;
+layout(location = 3) in float texID;
 
 out vec2 v_TexCoord;
-out int v_TexID;
+out float v_TexID;
 out vec4 v_Color;
 
 uniform mat4 u_MVP;
@@ -27,7 +27,7 @@ v_Color = color;
 #version 450 core
 
 in vec2 v_TexCoord;
-in int v_TexID;
+in float v_TexID;
 in vec4 v_Color;
 
 layout(location = 0) out vec4 color;
@@ -36,6 +36,6 @@ uniform sampler2D u_Texture [3];
 
 void main()
 {
-vec4 texColor = texture(u_Texture[v_TexID], v_TexCoord);
+vec4 texColor = texture(u_Texture[int(v_TexID)], v_TexCoord);
 color = mix(texColor, v_Color, 0.5);
 };
